fix: move helpers to toolkit
