+++ src/api/routes/ws/channels/ws-rpc-channel.ts
@@ -1,7 +1,7 @@
 import * as http from 'http';
 import * as WebSocket from 'ws';
 import * as net from 'net';
-import { isValidPrincipal, normalizeHashString, resolveOrTimeout } from '../../../../helpers';
+import { isValidPrincipal, normalizeHashString } from '../../../../helpers';
 import { WebSocketPrometheus } from '../web-socket-prometheus';
 import {
   ListenerType,
@@ -38,7 +38,7 @@ import {
 } from '@stacks/stacks-blockchain-api-types';
 import { getWsMessageTimeoutMs, getWsPingIntervalMs } from '../web-socket-transmitter';
 import { logger } from '../../../../logger';
-import { isProdEnv } from '@hirosystems/api-toolkit';
+import { isProdEnv, resolveOrTimeout } from '@hirosystems/api-toolkit';
 
 type Subscription =
   | RpcTxUpdateSubscriptionParams
