+++ src/ec-helpers.ts
@@ -1,7 +1,7 @@
 import * as bitcoin from 'bitcoinjs-lib';
 import * as ecc from 'tiny-secp256k1';
 import { ECPairAPI, ECPairFactory, ECPairInterface } from 'ecpair';
-import { coerceToBuffer } from './helpers';
+import { coerceToBuffer } from '@hirosystems/api-toolkit';
 
 export { ECPairInterface };
