+++ src/event-replay/connection-legacy.ts
@@ -1,11 +1,58 @@
+// TODO: This file exists because we use the old `pg` library to stream node events during replays.
+// we should migrate replays into `postgres.js` and delete this file.
+
 import { Client, ClientConfig, Pool, PoolClient, PoolConfig } from 'pg';
-import { parseArgBoolean, parsePort, stopwatch, timeout } from '../helpers';
 import { PgServer } from '../datastore/connection';
-import { isPgConnectionError } from '../datastore/helpers';
 import { logger } from '../logger';
+import { parseBoolean, stopwatch, timeout } from '@hirosystems/api-toolkit';
+import { parsePort } from '../helpers';
 
 export type PgClientConfig = ClientConfig & { schema?: string };
 type PgPoolConfig = PoolConfig & { schema?: string };
+
+/**
+ * Checks if a given error from the pg lib is a connection error (i.e. the query is retryable).
+ * If true then returns a normalized error message, otherwise returns false.
+ */
+export function isPgConnectionError(error: any): string | false {
+  if (error.code === 'ECONNREFUSED') {
+    return 'Postgres connection ECONNREFUSED';
+  } else if (error.code === 'ETIMEDOUT') {
+    return 'Postgres connection ETIMEDOUT';
+  } else if (error.code === 'ENOTFOUND') {
+    return 'Postgres connection ENOTFOUND';
+  } else if (error.code === 'ECONNRESET') {
+    return 'Postgres connection ECONNRESET';
+  } else if (error.code === 'CONNECTION_CLOSED') {
+    return 'Postgres connection CONNECTION_CLOSED';
+  } else if (error.code === 'CONNECTION_ENDED') {
+    return 'Postgres connection CONNECTION_ENDED';
+  } else if (error.code === 'CONNECTION_DESTROYED') {
+    return 'Postgres connection CONNECTION_DESTROYED';
+  } else if (error.code === 'CONNECTION_CONNECT_TIMEOUT') {
+    return 'Postgres connection CONNECTION_CONNECT_TIMEOUT';
+  } else if (error.code === 'CONNECT_TIMEOUT') {
+    return 'Postgres connection CONNECT_TIMEOUT';
+  } else if (error.message) {
+    const msg = (error as Error).message.toLowerCase();
+    if (msg.includes('database system is starting up')) {
+      return 'Postgres connection failed while database system is starting up';
+    } else if (msg.includes('database system is shutting down')) {
+      return 'Postgres connection failed while database system is shutting down';
+    } else if (msg.includes('connection terminated unexpectedly')) {
+      return 'Postgres connection terminated unexpectedly';
+    } else if (msg.includes('connection terminated')) {
+      return 'Postgres connection terminated';
+    } else if (msg.includes('connection error')) {
+      return 'Postgres client has encountered a connection error and is not queryable';
+    } else if (msg.includes('terminating connection due to unexpected postmaster exit')) {
+      return 'Postgres connection terminating due to unexpected postmaster exit';
+    } else if (msg.includes('getaddrinfo eai_again')) {
+      return 'Postgres connection failed due to a DNS lookup error';
+    }
+  }
+  return false;
+}
 
 /**
  * Connects to a Postgres pool. This function will also test the connection first to make sure
@@ -131,7 +178,7 @@ export function getPgClientConfig<TGetPoolConfig extends boolean = false>({
       password: pgEnvVars.password,
       host: pgEnvVars.host,
       port: parsePort(pgEnvVars.port),
-      ssl: parseArgBoolean(pgEnvVars.ssl),
+      ssl: parseBoolean(pgEnvVars.ssl),
       schema: pgEnvVars.schema,
       application_name: appName,
     };
