+++ src/test-utils/shared-setup.ts
@@ -1,5 +1,5 @@
 import { StacksCoreRpcClient } from '../core-rpc/client';
-import { loadDotEnv, timeout } from '../helpers';
+import { loadDotEnv } from '../helpers';
 import { PgWriteStore } from '../datastore/pg-write-store';
 import {
   DummyEventMessageHandler,
@@ -9,6 +9,7 @@ import {
 import { ChainID } from '@stacks/common';
 import * as isCI from 'is-ci';
 import { migrate } from './test-helpers';
+import { timeout } from '@hirosystems/api-toolkit';
 
 interface GlobalTestEnv {
   db: PgWriteStore;
